pyproject.toml
[project]
name = "contextuse"
version = "0.1.0"
description = "Configurable ETL library for processing data archives (ChatGPT, Instagram) into structured threads"
authors = [
    { name = "onfabric", email = "dev@onfabric.io" },
]
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "ijson>=3.0",
]

[project.optional-dependencies]
disk = []  # no extra deps needed, uses stdlib pathlib
gcs = ["google-cloud-storage>=2.10.0,<3"]
postgres = ["psycopg2-binary>=2.9"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contextuse"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = true

